refactor(simd): split imgproc SIMD helpers into src/imgproc/simd.rs
Move 5 imgproc-specific SIMD functions (simd_rgb_to_gray_u8,
simd_bgr_to_gray_u8, simd_rgba_to_gray_u8, simd_bgra_to_gray_u8,
simd_deriv_3x3_row_f32) from src/core/simd.rs to a new
src/imgproc/simd.rs module for better separation of concerns.

Adds unit tests for bgr, rgba, bgra color conversion and deriv_3x3
SIMD kernels that were previously untested directly.

Closes #17
